Renamed integrators to be more explicit, started separating them
They still retain their order in the name.
