+++ src/lib.rs
@@ -1,9 +1,1 @@
 pub mod propagators;
-
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        assert_eq!(2 + 2, 4);
-    }
-}
